address comments, add tests
